requirements.txt
python-dotenv
openai
opencv-python
whisper-mic
resemble
PyAudio
numpy
requests